Moved ps functions
